Fix migrations, move crawler from model to service
