--- db/migrate/20161120030344_add_url_to_script.rb
@@ -1,5 +1,0 @@
-class AddUrlToScript < ActiveRecord::Migration
-  def change
-    add_column :scripts, :url, :string
-  end
-end
